Restructure KVM agent code
