+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/ha/KvmHaChecker.java
@@ -1,4 +1,4 @@
-package com.cloud.agent.resource.kvm;
+package com.cloud.agent.resource.kvm.ha;
 
 import com.cloud.utils.script.OutputInterpreter;
 import com.cloud.utils.script.Script;
