+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/vif/VifDriverBase.java
@@ -1,5 +1,7 @@
-package com.cloud.agent.resource.kvm;
+package com.cloud.agent.resource.kvm.vif;
 
+import com.cloud.agent.resource.kvm.LibvirtComputingResource;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef;
 import com.cloud.legacymodel.exceptions.InternalErrorException;
 import com.cloud.legacymodel.to.NicTO;
 import com.cloud.model.enumeration.NicModel;
