+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/wrapper/LibvirtPlugNicCommandWrapper.java
@@ -1,8 +1,8 @@
 package com.cloud.agent.resource.kvm.wrapper;
 
 import com.cloud.agent.resource.kvm.LibvirtComputingResource;
-import com.cloud.agent.resource.kvm.LibvirtVmDef.InterfaceDef;
-import com.cloud.agent.resource.kvm.VifDriver;
+import com.cloud.agent.resource.kvm.vif.VifDriver;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef;
 import com.cloud.common.request.ResourceWrapper;
 import com.cloud.legacymodel.communication.answer.Answer;
 import com.cloud.legacymodel.communication.answer.PlugNicAnswer;
@@ -34,9 +34,9 @@ public final class LibvirtPlugNicCommandWrapper
             final Connect conn = libvirtUtilitiesHelper.getConnectionByVmName(vmName);
             vm = libvirtComputingResource.getDomain(conn, vmName);
 
-            final List<InterfaceDef> pluggedNics = libvirtComputingResource.getInterfaces(conn, vmName);
+            final List<LibvirtVmDef.InterfaceDef> pluggedNics = libvirtComputingResource.getInterfaces(conn, vmName);
             Integer nicnum = 0;
-            for (final InterfaceDef pluggedNic : pluggedNics) {
+            for (final LibvirtVmDef.InterfaceDef pluggedNic : pluggedNics) {
                 if (pluggedNic.getMacAddress().equalsIgnoreCase(nic.getMac())) {
                     s_logger.debug("found existing nic for mac " + pluggedNic.getMacAddress() + " at index " + nicnum);
                     return new PlugNicAnswer(command, true, "success");
@@ -44,7 +44,7 @@ public final class LibvirtPlugNicCommandWrapper
                 nicnum++;
             }
             final VifDriver vifDriver = libvirtComputingResource.getVifDriver(nic.getType());
-            final InterfaceDef interfaceDef = vifDriver.plug(nic, "Default - VirtIO capable OS (64-bit)", "");
+            final LibvirtVmDef.InterfaceDef interfaceDef = vifDriver.plug(nic, "Default - VirtIO capable OS (64-bit)", "");
             vm.attachDevice(interfaceDef.toString());
 
             return new PlugNicAnswer(command, true, "success");
