+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/wrapper/LibvirtStopCommandWrapper.java
@@ -1,9 +1,9 @@
 package com.cloud.agent.resource.kvm.wrapper;
 
 import com.cloud.agent.resource.kvm.LibvirtComputingResource;
-import com.cloud.agent.resource.kvm.LibvirtVmDef.InterfaceDef;
-import com.cloud.agent.resource.kvm.VifDriver;
+import com.cloud.agent.resource.kvm.vif.VifDriver;
 import com.cloud.agent.resource.kvm.xml.LibvirtDiskDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef;
 import com.cloud.common.request.ResourceWrapper;
 import com.cloud.legacymodel.communication.answer.Answer;
 import com.cloud.legacymodel.communication.answer.StopAnswer;
@@ -45,13 +45,13 @@ public final class LibvirtStopCommandWrapper extends LibvirtCommandWrapper<StopC
             final Connect conn = libvirtUtilitiesHelper.getConnectionByVmName(vmName);
 
             final List<LibvirtDiskDef> disks = libvirtComputingResource.getDisks(conn, vmName);
-            final List<InterfaceDef> ifaces = libvirtComputingResource.getInterfaces(conn, vmName);
+            final List<LibvirtVmDef.InterfaceDef> ifaces = libvirtComputingResource.getInterfaces(conn, vmName);
             final String result = libvirtComputingResource.stopVm(conn, vmName, command.isForceStop());
             if (result == null) {
                 for (final LibvirtDiskDef disk : disks) {
                     libvirtComputingResource.cleanupDisk(disk);
                 }
-                for (final InterfaceDef iface : ifaces) {
+                for (final LibvirtVmDef.InterfaceDef iface : ifaces) {
                     // We don't know which "traffic type" is associated with
                     // each interface at this point, so inform all vif drivers
                     for (final VifDriver vifDriver : libvirtComputingResource.getAllVifDrivers()) {
