+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/xml/LibvirtCapXmlParser.java
@@ -1,4 +1,4 @@
-package com.cloud.agent.resource.kvm;
+package com.cloud.agent.resource.kvm.xml;
 
 import java.io.IOException;
 import java.io.StringReader;
