+++ cosmic-agent/src/main/java/com/cloud/agent/resource/kvm/xml/LibvirtDomainXmlParser.java
@@ -1,9 +1,8 @@
-package com.cloud.agent.resource.kvm;
+package com.cloud.agent.resource.kvm.xml;
 
-import com.cloud.agent.resource.kvm.LibvirtVmDef.InterfaceDef;
-import com.cloud.agent.resource.kvm.LibvirtVmDef.RngDef;
-import com.cloud.agent.resource.kvm.LibvirtVmDef.WatchDogDef;
-import com.cloud.agent.resource.kvm.xml.LibvirtDiskDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef.InterfaceDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef.RngDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef.WatchDogDef;
 import com.cloud.model.enumeration.DiskControllerType;
 import com.cloud.model.enumeration.NicModel;
 import com.cloud.model.enumeration.RngBackendModel;
