+++ cosmic-agent/src/test/java/com/cloud/agent/resource/kvm/LibvirtComputingResourceTest.java
@@ -10,14 +10,17 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import com.cloud.agent.resource.kvm.KvmHaBase.NfsStoragePool;
-import com.cloud.agent.resource.kvm.LibvirtVmDef.InterfaceDef;
+import com.cloud.agent.resource.kvm.ha.KvmHaBase.NfsStoragePool;
+import com.cloud.agent.resource.kvm.ha.KvmHaMonitor;
 import com.cloud.agent.resource.kvm.storage.KvmPhysicalDisk;
 import com.cloud.agent.resource.kvm.storage.KvmStoragePool;
 import com.cloud.agent.resource.kvm.storage.KvmStoragePoolManager;
+import com.cloud.agent.resource.kvm.vif.VifDriver;
 import com.cloud.agent.resource.kvm.wrapper.LibvirtRequestWrapper;
 import com.cloud.agent.resource.kvm.wrapper.LibvirtUtilitiesHelper;
 import com.cloud.agent.resource.kvm.xml.LibvirtDiskDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef;
+import com.cloud.agent.resource.kvm.xml.LibvirtVmDef.InterfaceDef;
 import com.cloud.common.storageprocessor.Processor;
 import com.cloud.common.storageprocessor.TemplateLocation;
 import com.cloud.common.storageprocessor.resource.StorageSubsystemCommandHandler;
